Update Examples to work with GxEPD 3.0+
This commit updates the example code to work with the latest version of GxEPD.

As shown in the [commit](https://github.com/ZinggJM/GxEPD/commit/e2869266d03eddef1d96eaab63b0a6f75eb7f91f), the GxGDEH029A1 files were not changed but the method for referencing / including them was.

It also adds the undocumented library requirements on the example readme.md for Weather (Time) and WebSocketDraw (WebSockets).