Clarity and proper imperial units
The weather example does a deep sleep which requires old-school restart before upload in order to work.  This wasn't clear especially given the prior version of the code.

Also, I've added full support for imperial units because Americans are lame.

Finally, I added the pinout image to the hardware folder for easy reference for others.